fix starting line number

+++ src/main/java/de/siegmar/fastcsv/reader/RowReader.java
@@ -247,7 +247,7 @@ final class RowReader {
         return shift;
     }
 
-    void resetBuffer(final int originalLineNumber) {
+    void resetBuffer(final long originalLineNumber) {
         rowHandler.setOriginalLineNumber(originalLineNumber);
         buffer.reset();
     }
